essai 02

+++ ProjetAlex/Program.cs
@@ -3,7 +3,7 @@ Console.WriteLine("Merci de m'indiquer un nombre pour définir la taille du tabl
 string nbsaisi01 = Console.ReadLine();
 
 // on transforme la saisie string en nombre entier et on l'affecte à nb01
-bool succes01 = int.TryParse(nbsaisi01, out decimal nb01);
+bool succes01 = int.TryParse(nbsaisi01, out int nb01);
 
 
 
@@ -12,7 +12,7 @@ if (succes01 == true)
 {
     int[] tableau2 = new int[nb01];
 
-for (int i=0; i<nb01; i++)
+for (int i=0; i < nb01; i++)
 {
     tableau2[i] = i;
 }
